feat: 블랙잭일 경우, 베팅 금액의 1.5배를 받는 기능 추가

+++ src/main/java/blackjack/domain/card/Result.java
@@ -2,5 +2,5 @@ package blackjack.domain.card;
 
 import java.util.Map;
 
-public record Result(Map<WinningResult, Integer> resultIntegerMap, int bettingResult) {
+public record Result(Map<WinningResult, Integer> resultIntegerMap, double bettingResult) {
 }
